# api/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Grocery & Meal Planner API
  description: |
    MVP API for managing pantry inventory, grocery lists, meal ideas, and expiration notifications.
    
    ## Features
    - **Inventory Management**: CRUD operations for pantry items with low-stock and expiring-soon flags
    - **Grocery Lists**: Create and manage shopping lists, auto-add low stock or expiring items
    - **Meal Ideas**: Get suggestions based on pantry contents, save favorite meals
    - **Notifications**: Schedule reminders for expiring items
  version: 1.0.0
  contact:
    name: Smart Grocery Team
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/meals/suggest:
    post:
      tags: [Meals]
      summary: Get AI meal suggestions based on pantry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
                additionalPreferences:
                  type: string
      responses:
        '200':
          description: Meal suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealSuggestionsResponse'
  /api/meals:
    get:
      tags: [Meals]
      summary: List saved meal ideas
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of saved meals
    post:
      tags: [Meals]
      summary: Save a meal idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMealIdea'
      responses:
        '201':
          description: Meal saved
  /api/meals/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Meals]
      summary: Get meal idea
      responses:
        '200':
          description: Meal idea
    put:
      tags: [Meals]
      summary: Update meal idea
      responses:
        '200':
          description: Meal updated
    delete:
      tags: [Meals]
      summary: Delete meal idea
      responses:
        '200':
          description: Meal deleted

  # Notifications
...
components:
  schemas:
...
    CreateMealIdea:
      type: object
      required: [userId, title]
      properties:
        userId:
          type: integer
        title:
          type: string
          maxLength: 255
        notes:
          type: string
          maxLength: 2000
    MealSuggestionsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            suggestions:
              type: array
              items:
                type: object
                properties:
                  title:
                    type: string
                  ingredients:
                    type: array
                    items:
                      type: string
                  instructions:
                    type: string

    # Notification schemas
...